Refactor: additional test fixups
Added openapi validation logs
Fixup backend director setup
Fixup backend overall timeout
